Reworking Theme selection, code formatting
Themes now are calculated with the theme-loader component. Which listens for changes in the <html> tag, in
this case changing the data-mode attribute. When a change happens (theme) a cookie is created with the new
theme. In case of system theme, the theme is calculated on the client side, otherwhise it's server side styled.
The theme select is a dumb component that only changes the data-mode attribute.